+++ src/main/java/org/sopt/app/application/app_service/AppServiceService.java
@@ -3,6 +3,7 @@ package org.sopt.app.application.app_service;
 import java.util.Comparator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.sopt.app.application.app_service.dto.AppServiceInfo;
 import org.sopt.app.domain.entity.AppService;
 import org.sopt.app.interfaces.postgres.AppServiceRepository;
 import org.springframework.stereotype.Service;
