+++ src/main/java/org/sopt/app/application/app_service/dto/AppServiceBadgeInfo.java
@@ -1,0 +1,7 @@
+package org.sopt.app.application.app_service.dto;
+
+public record AppServiceBadgeInfo(
+
+) {
+
+}
